make movie carousel a component, remove unwanted elements
